Refactor kubeproxy unit tests to improve maintainability
- Move testDriver to kubeproxy_suite_test.go
- Move testEndpoint() to endpoint_test.go
- Move testGatewayTransition() to gw_transition_test.go
