+++ pkg/routeagent_driver/handlers/kubeproxy/sync_handler_test.go
@@ -19,39 +19,11 @@ limitations under the License.
 package kubeproxy_test
 
 import (
-	"net"
-	"time"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	submarinerv1 "github.com/submariner-io/submariner/pkg/apis/submariner.io/v1"
-	"github.com/submariner-io/submariner/pkg/cni"
-	"github.com/submariner-io/submariner/pkg/event/testing"
-	netlinkAPI "github.com/submariner-io/submariner/pkg/netlink"
-	fakeNetlink "github.com/submariner-io/submariner/pkg/netlink/fake"
-	"github.com/submariner-io/submariner/pkg/packetfilter"
-	fakePF "github.com/submariner-io/submariner/pkg/packetfilter/fake"
 	"github.com/submariner-io/submariner/pkg/routeagent_driver/constants"
 	"github.com/submariner-io/submariner/pkg/routeagent_driver/handlers/kubeproxy"
-	"github.com/vishvananda/netlink"
-	"golang.org/x/sys/unix"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/uuid"
-	k8snet "k8s.io/utils/net"
-)
-
-const (
-	cniIPAddress     = "192.168.5.1"
-	localClusterCIDR = cniIPAddress + "/24"
-	localServiceCIDR = "169.254.2.0/24"
-	remoteSubnet1    = "170.250.1.0/24"
-	remoteSubnet2    = "171.250.1.0/24"
-	localNodeName1   = "local-node1"
-	localNodeName2   = "local-node2"
-	remoteNodeName   = "remote-node"
-	nodeAddress1     = "10.253.10.2"
-	nodeAddress2     = "10.253.10.3"
 )
 
 var _ = Describe("SyncHandler", func() {
@@ -60,246 +32,6 @@ var _ = Describe("SyncHandler", func() {
 	Describe("Nodes", testNodes)
 	Describe("Uninstall", testUninstall)
 })
-
-func testEndpoints() {
-	t := newTestDriver()
-
-	When("a local Endpoint is created while on a non-gateway node", func() {
-		JustBeforeEach(func() {
-			t.CreateEndpoint(t.localEndpoint)
-		})
-
-		It("should add the VxLAN interface", func() {
-			Expect(toVxlan(t.netLink.AwaitLink(kubeproxy.VxLANIface)).Group.String()).To(Equal(
-				t.localEndpoint.Spec.GetPrivateIP(k8snet.IPv4)))
-		})
-
-		Context("and old VxLAN routes are present", func() {
-			BeforeEach(func() {
-				t.addVxLANRoute(remoteSubnet1)
-			})
-
-			It("should remove them", func() {
-				t.netLink.AwaitNoDstRoutes(t.vxLanInterfaceIndex, 0, remoteSubnet1)
-			})
-		})
-
-		Context("and a VxLAN interface from a previous Endpoint exists", func() {
-			JustBeforeEach(func() {
-				t.netLink.AwaitLink(kubeproxy.VxLANIface)
-			})
-
-			It("should remove the previous VxLAN interface", func() {
-				t.netLink.SetLinkIndex(kubeproxy.VxLANIface, t.vxLanInterfaceIndex+1)
-				t.CreateEndpoint(newLocalEndpoint(localNodeName2))
-
-				Eventually(func() int {
-					return t.netLink.AwaitLink(kubeproxy.VxLANIface).Attrs().Index
-				}).Should(Equal(t.vxLanInterfaceIndex + 1))
-			})
-		})
-
-		Context("and remote subnets are present", func() {
-			BeforeEach(func() {
-				t.CreateEndpoint(t.remoteEndpoint)
-			})
-
-			It("should add VxLAN routes for the remote subnets", func() {
-				t.verifyVxLANRoutes()
-			})
-		})
-	})
-
-	When("a local Endpoint is removed while on a non-gateway node", func() {
-		BeforeEach(func() {
-			t.CreateEndpoint(t.localEndpoint)
-			t.netLink.AwaitLink(kubeproxy.VxLANIface)
-		})
-
-		JustBeforeEach(func() {
-			t.DeleteEndpoint(t.localEndpoint.Name)
-		})
-
-		Context("and its host name matches that associated with the existing VxLAN interface", func() {
-			It("should remove the existing VxLAN interface", func() {
-				t.netLink.AwaitNoLink(kubeproxy.VxLANIface)
-			})
-		})
-
-		Context("and its host name does not match that associated with the existing VxLAN interface", func() {
-			BeforeEach(func() {
-				t.localEndpoint.Spec.Hostname = localNodeName2
-			})
-
-			It("should not remove the existing VxLAN interface", func() {
-				t.netLink.AwaitLink(kubeproxy.VxLANIface)
-			})
-		})
-
-		Context("and is subsequently recreated", func() {
-			It("should recreate the VxLAN interface", func() {
-				t.netLink.AwaitNoLink(kubeproxy.VxLANIface)
-
-				t.CreateEndpoint(t.localEndpoint)
-				t.netLink.AwaitLink(kubeproxy.VxLANIface)
-			})
-		})
-	})
-
-	When("a local Endpoint is created while on a gateway node", func() {
-		It("should not add the VxLAN interface", func() {
-			t.localEndpoint.Spec.Hostname = t.Hostname
-			t.CreateEndpoint(t.localEndpoint)
-			t.netLink.AwaitNoLink(kubeproxy.VxLANIface)
-		})
-	})
-
-	When("a remote Endpoint is created while on a non-gateway node", func() {
-		JustBeforeEach(func() {
-			t.CreateEndpoint(t.remoteEndpoint)
-		})
-
-		Context("after a local Endpoint was created", func() {
-			BeforeEach(func() {
-				t.CreateEndpoint(t.localEndpoint)
-			})
-
-			It("should add VxLAN routes for the remote subnets", func() {
-				t.verifyVxLANRoutes()
-			})
-
-			It("should add IP table rules for the remote subnets", func() {
-				t.verifyRemoteSubnetIPTableRules()
-			})
-
-			It("should not add routing rules for host networking", func() {
-				t.verifyNoHostNetworkingRoutes()
-			})
-
-			Context("and is subsequently removed", func() {
-				JustBeforeEach(func() {
-					t.DeleteEndpoint(t.remoteEndpoint.Name)
-				})
-
-				It("should remove the VxLAN routes for the remote subnets", func() {
-					t.verifyNoVxLANRoutes()
-				})
-			})
-		})
-
-		Context("before a local Endpoint is created", func() {
-			It("should not add VxLAN routes for the remote subnets", func() {
-				t.verifyNoVxLANRoutes()
-			})
-
-			It("should add IP table rules for the remote subnets", func() {
-				t.verifyRemoteSubnetIPTableRules()
-			})
-		})
-
-		Context("and is subsequently removed followed by a local Endpoint created", func() {
-			JustBeforeEach(func() {
-				t.DeleteEndpoint(t.remoteEndpoint.Name)
-				t.CreateEndpoint(t.localEndpoint)
-			})
-
-			It("should not add VxLAN routes for the remote subnets", func() {
-				t.verifyNoVxLANRoutes()
-			})
-		})
-	})
-
-	When("a remote Endpoint is created while on a gateway node", func() {
-		JustBeforeEach(func() {
-			t.CreateLocalHostEndpoint()
-			t.CreateEndpoint(t.remoteEndpoint)
-		})
-
-		It("should not add VxLAN routes for the remote subnets", func() {
-			t.verifyNoVxLANRoutes()
-		})
-
-		It("should add IP table rules for the remote subnets", func() {
-			t.verifyRemoteSubnetIPTableRules()
-		})
-
-		It("should add routing rules for host networking", func() {
-			t.verifyHostNetworkingRoutes()
-		})
-
-		Context("and is subsequently removed", func() {
-			JustBeforeEach(func() {
-				t.DeleteEndpoint(t.remoteEndpoint.Name)
-			})
-
-			It("should remove routing rules for host networking", func() {
-				t.verifyNoHostNetworkingRoutes()
-			})
-		})
-	})
-}
-
-func testGatewayTransition() {
-	t := newTestDriver()
-
-	When("transition to gateway", func() {
-		var localHostEP *submarinerv1.Endpoint
-
-		JustBeforeEach(func() {
-			t.CreateEndpoint(t.remoteEndpoint)
-			localHostEP = t.CreateLocalHostEndpoint()
-		})
-
-		It("should add the VxLAN interface", func() {
-			Expect(toVxlan(t.netLink.AwaitLink(kubeproxy.VxLANIface)).Group).To(BeNil())
-		})
-
-		It("should add a routing rule for the RouteAgentHostNetworkTableID", func() {
-			t.netLink.AwaitRule(constants.RouteAgentHostNetworkTableID, "", "")
-		})
-
-		It("should add host networking routing rules for the remote subnets", func() {
-			t.verifyHostNetworkingRoutes()
-		})
-
-		Context("and previous VxLAN routes are present", func() {
-			BeforeEach(func() {
-				t.addVxLANRoute(remoteSubnet1)
-				t.CreateEndpoint(t.localEndpoint)
-			})
-
-			It("should remove them", func() {
-				t.netLink.AwaitNoDstRoutes(t.vxLanInterfaceIndex, 0, remoteSubnet1)
-			})
-		})
-
-		Context("and Node addresses are present", func() {
-			BeforeEach(func() {
-				t.CreateNode(newNode(nodeAddress1))
-				t.CreateNode(newNode(nodeAddress2))
-			})
-
-			It("should add an FDB entry on the VxLAN interface for each address", func() {
-				t.netLink.AwaitNeighbors(t.vxLanInterfaceIndex, nodeAddress1, nodeAddress2)
-			})
-		})
-
-		Context("and then to non-gateway", func() {
-			JustBeforeEach(func() {
-				t.netLink.AwaitRule(constants.RouteAgentHostNetworkTableID, "", "")
-				t.DeleteEndpoint(localHostEP.Name)
-			})
-
-			It("should remove the routing rule for the RouteAgentHostNetworkTableID", func() {
-				t.netLink.AwaitNoRule(constants.RouteAgentHostNetworkTableID, "", "")
-			})
-
-			It("should remove host networking routing rules for the remote subnets", func() {
-				t.verifyNoHostNetworkingRoutes()
-			})
-		})
-	})
-}
 
 func testNodes() {
 	t := newTestDriver()
@@ -350,151 +82,4 @@ func testUninstall() {
 			t.verifyNoHostNetworkingRoutes()
 		})
 	})
-}
-
-type testDriver struct {
-	*testing.ControllerSupport
-	handler             *kubeproxy.SyncHandler
-	pFilter             *fakePF.PacketFilter
-	netLink             *fakeNetlink.NetLink
-	localEndpoint       *submarinerv1.Endpoint
-	remoteEndpoint      *submarinerv1.Endpoint
-	hostInterfaceIndex  int
-	vxLanInterfaceIndex int
-}
-
-func newTestDriver() *testDriver {
-	t := &testDriver{
-		ControllerSupport: testing.NewControllerSupport(),
-	}
-
-	BeforeEach(func() {
-		t.hostInterfaceIndex = 100
-		t.vxLanInterfaceIndex = 200
-
-		t.netLink = fakeNetlink.New()
-		t.netLink.SetLinkIndex(kubeproxy.VxLANIface, t.vxLanInterfaceIndex)
-
-		t.netLink.SetupDefaultGateway(k8snet.IPv4, net.Interface{
-			Index: t.hostInterfaceIndex,
-			MTU:   100,
-			Name:  "gw-intf",
-		}, &net.IPNet{IP: net.ParseIP("1.2.3.4"), Mask: net.CIDRMask(8, 32)})
-
-		netlinkAPI.NewFunc = func() netlinkAPI.Interface {
-			return t.netLink
-		}
-		t.pFilter = fakePF.New()
-
-		cni.HostInterfaces = func() ([]cni.HostInterface, error) {
-			return []cni.HostInterface{{
-				Name: "veth0",
-				Addr: cniIPAddress + "/24",
-			}}, nil
-		}
-
-		t.localEndpoint = newLocalEndpoint(localNodeName1)
-		t.remoteEndpoint = newRemoteEndpoint()
-
-		t.handler = kubeproxy.NewSyncHandler([]string{localClusterCIDR}, []string{localServiceCIDR})
-
-		t.Start(t.handler)
-	})
-
-	return t
-}
-
-func (t *testDriver) verifyVxLANRoutes() {
-	t.netLink.AwaitDstRoutes(t.netLink.AwaitLink(kubeproxy.VxLANIface).Attrs().Index, 0, t.remoteEndpoint.Spec.Subnets...)
-}
-
-func (t *testDriver) verifyNoVxLANRoutes() {
-	time.Sleep(200 * time.Millisecond)
-	t.netLink.AwaitNoDstRoutes(t.vxLanInterfaceIndex, 0, t.remoteEndpoint.Spec.Subnets...)
-}
-
-func (t *testDriver) verifyHostNetworkingRoutes() {
-	t.netLink.AwaitDstRoutes(t.hostInterfaceIndex, constants.RouteAgentHostNetworkTableID, t.remoteEndpoint.Spec.Subnets...)
-}
-
-func (t *testDriver) verifyNoHostNetworkingRoutes() {
-	time.Sleep(200 * time.Millisecond)
-	t.netLink.AwaitNoDstRoutes(t.hostInterfaceIndex, constants.RouteAgentHostNetworkTableID, t.remoteEndpoint.Spec.Subnets...)
-}
-
-func (t *testDriver) verifyRemoteSubnetIPTableRules() {
-	for _, remoteCIDR := range t.remoteEndpoint.Spec.Subnets {
-		t.pFilter.AwaitRule(packetfilter.TableTypeNAT, constants.SmPostRoutingChain,
-			And(ContainSubstring(localClusterCIDR), ContainSubstring(remoteCIDR)))
-	}
-}
-
-func (t *testDriver) addVxLANRoute(cidr string) {
-	_, dst, err := net.ParseCIDR(cidr)
-	Expect(err).To(Succeed())
-
-	_ = t.netLink.RouteAdd(&netlink.Route{
-		Dst:       dst,
-		Gw:        net.IPv4(11, 21, 31, 41),
-		Scope:     unix.RT_SCOPE_UNIVERSE,
-		LinkIndex: t.vxLanInterfaceIndex,
-		Protocol:  4,
-	})
-}
-
-func newLocalEndpoint(hostname string) *submarinerv1.Endpoint {
-	return &submarinerv1.Endpoint{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: string(uuid.NewUUID()),
-		},
-		Spec: submarinerv1.EndpointSpec{
-			CableName:  "submariner-cable-local-192-68-1-2",
-			ClusterID:  testing.LocalClusterID,
-			PrivateIPs: []string{"192.68.1.2"},
-			Hostname:   hostname,
-			Backend:    "libreswan",
-		},
-	}
-}
-
-func newRemoteEndpoint() *submarinerv1.Endpoint {
-	return &submarinerv1.Endpoint{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: string(uuid.NewUUID()),
-		},
-		Spec: submarinerv1.EndpointSpec{
-			CableName:  "submariner-cable-remote-192-68-1-2",
-			ClusterID:  "remote",
-			PrivateIPs: []string{"192.68.1.2"},
-			Hostname:   remoteNodeName,
-			Subnets:    []string{remoteSubnet1, remoteSubnet2},
-			Backend:    "libreswan",
-		},
-	}
-}
-
-func newNode(addr string) *corev1.Node {
-	return &corev1.Node{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: string(uuid.NewUUID()),
-		},
-		Status: corev1.NodeStatus{
-			Addresses: []corev1.NodeAddress{
-				{
-					Type: corev1.NodeExternalDNS,
-				},
-				{
-					Type:    corev1.NodeInternalIP,
-					Address: addr,
-				},
-			},
-		},
-	}
-}
-
-func toVxlan(link netlink.Link) *netlink.Vxlan {
-	vxLan, ok := link.(*netlink.Vxlan)
-	Expect(ok).To(BeTrue(), "Unexpected Link type: %T", link)
-
-	return vxLan
 }
